Fix some minor issues in the README
Was recently playing around with the extension and noticed a couple of things in the README that could be improved, especially since the README shows up in the extension description of VS Code. Take it or leave it, not trying to nitpick, I just like to help out if I can.

* Fix a small typo ("manuall" -> "manual")
* Switch "Pico-SDK" to the technically correcter "Pico SDK"
* Monospace `Cmakelists.txt`
* Link to the releases page, which sorts from latest->oldest, instead of a specific beta (that you need to always remember to update, it always eventually slips my mind)

Thanks for all the hard work on this extension!